requirements.txt
aiohttp==3.5.4
astroid==2.2.5
async-timeout==3.0.1
attrs==19.1.0
certifi==2019.6.16
chardet==3.0.4
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.1
mccabe==0.6.1
multidict==4.5.2
pylint==2.3.1
requests==2.22.0
six==1.12.0
slackclient==2.1.0
typed-ast==1.4.0
urllib3==1.25.3
wrapt==1.11.2
yarl==1.3.0
